Fixed notifications#notify to at least work
It does send out the notifications, but more work is needed, especially
around displaying the info. This commit only supports email, not sms.
This commit also creates a Message object to record the sending of the
message.
This commit also adds Inventory as a role in the CanCan authentication,
although the menus do not yet look for it.